Setting up opening screen to find planets and select "Start game"
